+++ src/main/java/com/fieldmanagement/fieldmanagement_be/field/domain/model/HourlyRate.java
@@ -21,5 +21,5 @@ public class HourlyRate extends BaseModel {
     String id;
     LocalTime startTime;
     LocalTime endTime;
-    Float price;
+    Long price;
 }
